power/power_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Power API</title>
</head>
<body>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

<script type="application/javascript">

function postMsg(msg)
{
  parent.postMessage(msg, "http://mochi.test:8888");
}


var result = {
  "result": 'mozPower' in navigator, 
  "msg": "navigator.mozPower Should exist"};


postMsg(result);

/** Test permission **/

var power = navigator.mozPower;

var result = {
"result": power ? true : false, 
"msg": "navigator.mozPower is enabled in browser"};

postMsg(result);


//parent.ok(!power, "Should be able to access power manager with permission.");

parent.postMessage("finished", "*");

</script>
</pre>
</body>
</html>
